refactor: 이전 작업한 뷰, 뷰컨 분리 및 코드수정
LearningViewController
-> LearningView, LearningViewController

LearningResultViewController
-> LearningResultView, LearningResultViewController

RecordViewController
-> RecordView, RecordViewController
