+++ VocaVocca/View/Learning/LearningViewController.swift
@@ -10,54 +10,32 @@ import SnapKit
 
 final class LearningViewController: UIViewController {
     
-    private let flowLayout: UICollectionViewFlowLayout = {
-        let layout = UICollectionViewFlowLayout()
-        layout.scrollDirection = .vertical
-        layout.minimumInteritemSpacing = 20
-        layout.minimumLineSpacing = 20
-        layout.sectionInset = UIEdgeInsets(top: 0, left: 16, bottom: 0, right: 16)
-        return layout
-    }()
+    private let learningView = LearningView()
     
-    private lazy var collectionView: UICollectionView = {
-        let collectionView = UICollectionView(frame: .zero, collectionViewLayout: flowLayout)
-        collectionView.backgroundColor = .white
-        collectionView.register(LearningViewCell.self, forCellWithReuseIdentifier: LearningViewCell.id)
-        collectionView.dataSource = self
-        collectionView.delegate = self
-        return collectionView
-    }()
-    
-    private lazy var startButton: CustomButton = {
-        let button = CustomButton(title: "시작하기", action: startButtonTapped)
-        return button
-    }()
+    override func loadView() {
+        self.view = learningView
+    }
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        setupUI()
+        setupCollectionView()
+        setupActions()
     }
     
-    private func setupUI() {
-        view.backgroundColor = .white
-        view.addSubviews(collectionView, startButton)
-        
-        collectionView.snp.makeConstraints {
-            $0.horizontalEdges.equalTo(view.safeAreaLayoutGuide)
-            $0.top.equalTo(view.safeAreaLayoutGuide.snp.top)
-            $0.bottom.equalTo(view.safeAreaLayoutGuide.snp.bottom).offset(-100)
-        }
-        
-        startButton.snp.makeConstraints {
-            $0.bottom.equalTo(view.safeAreaLayoutGuide.snp.bottom).offset(-20)
-            $0.horizontalEdges.equalToSuperview().inset(40)
-            $0.height.equalTo(50)
-            $0.centerX.equalToSuperview()
-        }
+    private func setupCollectionView() {
+        learningView.collectionView.dataSource = self
+        learningView.collectionView.delegate = self
+    }
+    
+    private func setupActions() {
+        learningView.startButton.addTarget(self, action: #selector(startButtonTapped), for: .touchUpInside)
     }
     
     @objc private func startButtonTapped() {
-        print(#function)
+        // 플래시카드뷰 띄우기
+        let flashcardVC = FlashcardViewController()
+        flashcardVC.modalPresentationStyle = .fullScreen
+        present(flashcardVC, animated: false)
     }
 }
 
@@ -76,14 +54,8 @@ extension LearningViewController: UICollectionViewDataSource {
     }
 }
 
-// MARK: - UICollectionViewDelegateFlowLayout
+// MARK: - UICollectionViewDelegate
 
-extension LearningViewController: UICollectionViewDelegateFlowLayout {
-    // 아이템 크기 설정
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
-        // 열 간격 + 좌우 inset 합
-        let totalSpacing = 20 + 16 * 2
-        let width = (Int(collectionView.frame.width) - totalSpacing) / 2
-        return CGSize(width: width, height: 100)
-    }
+extension LearningViewController: UICollectionViewDelegate {
+    
 }
